feat: Normalize IMU data and use buffers for IMU's
This PR does a major overhaul of how IMU data is managed across every
aspect of InputPlumber.
- All source device drivers now normalize IMU data into radians per second
  for gyroscope data and meters per second squared for accelerometer data.
- All target devices now denormalize Accelerometer and Gyroscope events
  into their native format (typically LSB).
- The iio_imu source device driver now configures a buffer and follows
  an asynchronous drain/refill pattern to create events. This reduces
  CPU and RAM overhead by reducing the number of reads per event from
  3-6 to 1. This also has the benefit of reducing jitter when using
  iio_imu based devices as the entire event is now based on a single
  timestamp, when the old method had significant drift during each poll.
- Removes deprecated Gamepad::Accelerometer and Gamepad:Gyro events.

+++ src/drivers/legos/event.rs
@@ -87,9 +87,9 @@ pub enum AxisEvent {
 /// [InertialInput] represents the state of the IMU (x, y, z) values
 #[derive(Clone, Debug)]
 pub struct InertialInput {
-    pub x: i16,
-    pub y: i16,
-    pub z: i16,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 /// [InertialEvent] has data from the IMU
